Typo: move UpcomingView into own folder.
