+++ lib/widgets/home_screen/home_screen.dart
@@ -4,9 +4,9 @@ import 'package:film_freund/services/service_locator.dart';
 import 'package:film_freund/utils/sizes.dart';
 import 'package:film_freund/widgets/home_screen/active_view.dart';
 import 'package:film_freund/widgets/home_screen/popular/popular_view.dart';
-import 'package:film_freund/widgets/home_screen/popular/upcoming_view.dart';
 import 'package:film_freund/widgets/home_screen/settings/settings_view.dart';
 import 'package:film_freund/widgets/home_screen/sidebar.dart';
+import 'package:film_freund/widgets/home_screen/upcoming/upcoming_view.dart';
 import 'package:film_freund/widgets/signin_screen/signin_screen.dart';
 import 'package:flutter/material.dart';
