+++ test/widgets/home_screen/home_screen_test.dart
@@ -6,8 +6,8 @@ import 'package:film_freund/utils/sizes.dart';
 import 'package:film_freund/widgets/home_screen/active_view.dart';
 import 'package:film_freund/widgets/home_screen/home_screen.dart';
 import 'package:film_freund/widgets/home_screen/popular/popular_view.dart';
-import 'package:film_freund/widgets/home_screen/popular/upcoming_view.dart';
 import 'package:film_freund/widgets/home_screen/settings/settings_view.dart';
+import 'package:film_freund/widgets/home_screen/upcoming/upcoming_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
